Began BinaryTree implementation
I have been held up on getting this project started
for too long. I wanted to start off with the perfect
plan and make sure everything is ready to be generalized.
But, I just never could start with all the planning.
I am just getting stuff down now and I can refactor it
at a later point to work in a hierarchy.